Fix "View on GitHub" links (#1662)
* separated controlExamples.tsx into individual files: radioExample.tsx, checkboxExample.tsx, and switchExample.tsx

* formatting fix: added newline to end of files

* removed unnecessary tslint disabling
